api/requirements.txt
fastapi
uvicorn
psycopg[binary]
python-jose
passlib[bcrypt]